Fixed issue 13
Issue: https://github.com/mikekelly/hal-browser/issues/13

The dialog that appear when trying to GET a templated URI is now prepopulated
with JSON that contains the expressions available in the URI template.

It contains all expressions regardless whether they are optional or mandatory.

The implementation is tightly coupled with the current implementation of
uritemplate.js.