feature: Provide support for server authentication
Provide capability to specify username and password, and pass those credentials on in the Ajax request. Enable user
to turn credentials on and off.

NB: Server password is stored in memory as cleartext. Credentials should only be used for testing purposes, and not
in a production environment with significant security constraints.