Add support for following relative _links
This adds the new vendored JS library URI.js into the web page.  But it also includes
an implementation of normalizeUrl in the HAL object which can then be called by the
various bits in browser.html to normalize URIs.  Existing absolute URIs are not
affected by this change.  Only if an href is a relative URL is the URI normalized and
it is normalized with respect to the current resource being inspected by the HAL
browser (NOT the URL of the current web page).

Note that a web browser does this automatically.  In fact, it even does it for the
_links (e.g., if you mouse over the GET buttons).  The problem is that it normalizes
with respect to the wrong Base URI.  All this does is give the HAL browser the same
capability that a normal browser has...to identify the absolute URI of a resource
given a relative URI and a context.