Add extension point for the non-GET dialog
This patch provides the means to inject a different non-GET dialog. The default dialog requires useres to type in a JSON document for POST support, requiring the user to know what is required. By offering an extension, developers can code a more intelligent pop-up based upon alternative metadata not supplied by this module.

After this is merged, and I can use this in my own project, I am happy to update the documentation with a polished example of using this extension.

Resolves #63